order

+++ src/Jhipster.Module/OrderSvc/OrderSvc.Application/Command/TransactionsCommand/AddTransactionsCommand.cs
@@ -19,7 +19,7 @@ namespace OrderSvc.Application.Command.TransactionsCommand
         public string? TransactionName { get; set; }
 
         public string TransactionType { get; set; }
-		public decimal? TotalAmount { get; set; }
+		public double? TotalAmount { get; set; }
 		public DateTime? TransactionDate { get; set; }
 		public string? PaymentMethod { get; set; }
 	}
